\name{cacheSweave}
\alias{cacheSweave}

\title{Cache Sweave Computations}
\description{
  Cache computations when using Sweave
}
\usage{
cacheSweave(prefix, expr, envir = parent.frame(), keys = NULL)
}

\arguments{
  \item{prefix}{character, to be used for naming the cache file}
  \item{expr}{code to be evaluated}
  \item{envir}{environment into which cached results should be loaded}
  \item{keys}{character vector of keys to be loaded}
}
\details{
  
}
\value{
  Nothing useful is returned.
}
\author{Roger D. Peng \email{rpeng@jhsph.edu}}

\keyword{utilities}% at least one, from doc/KEYWORDS

