\name{cacheSweave}
\alias{cacheSweaveDriver}

\title{Cache Sweave Computations}
\description{
  Cache computations when using Sweave
}
\usage{
cacheSweaveDriver()
}

\details{
  \code{cacheSweaveDriver} is used directly with \code{Sweave} and is
  passed as the \code{driver} argument.  \code{cacheSweave} can be used
  within the R code chunk.
}
\value{
  Nothing useful is returned.
}

\author{Roger D. Peng \email{rpeng@jhsph.edu}, with much code for
\code{cacheSweaveDriver} copied from Sweave source by Friedrich Leisch
}

\seealso{
  \code{\link{Sweave}}
}

\keyword{utilities}% at least one, from doc/KEYWORDS

